completed refactor to two components, mapContainer and geoMap
